\name{Est.KM.FUN.weighted}
\alias{Est.KM.FUN.weighted}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Nonparametric Nelson-Aalen estimate of survival}
\description{
Nonparametric Nelson-Aalen estimate of survival}
\usage{
Est.KM.FUN.weighted(xi, di, si, myt, weight.perturb = NULL, bw = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{xi}{
xi
}
  \item{di}{
di
}
  \item{si}{
si
}
  \item{myt}{
myt
}
  \item{weight.perturb}{
weight.perturb
}
  \item{bw}{
bw
}
}

\value{
Smoothed survival estimate.
}

\author{
Layla Parast
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{survival}
\keyword{internal}
